requirements.txt
huggingface-hub
sentence-transformers
faiss-cpu
streamlit
requests
lxml
numpy